separated default monitoring implementation in its own file
